refactor: init of canvas editor
